requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
Flask-Limiter==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
limits==2.0.3
MarkupSafe==2.0.1
pycodestyle==2.8.0
python-dotenv==0.19.2
python-http-client==3.3.4
redis==4.0.2
requests==2.26.0
sendgrid==6.9.2
starkbank-ecdsa==2.0.3
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
